feat(core): Add support for aliasing values
Reworked value parsing to streamline things and DRY it out a bit.

Fixes #719
Fixes #736
